Revert commits

+++ src/box_config/box_config.service.ts
@@ -105,12 +105,13 @@ export class BoxConfigService implements OnModuleInit {
       );
     }
     //TODO:comment in
-    // const isVerified = checkIfMessageIsSigned(
-    //   signedMessage,
-    //   'Update Primes Mint',
-    //   authority,
-    // );
-    // if(!isVerified) throw new UnauthorizedException()
+
+    const isVerified = checkIfMessageIsSigned(
+      signedMessage,
+      'Update Primes Mint',
+      authority,
+    );
+    if (!isVerified) throw new UnauthorizedException();
     const saved = await this.saveOrUpdateBox.execute(box);
     this.logger.debug(`Staring box worker with id:${saved.boxId}`);
     if (!box.boxId) {
@@ -151,12 +152,12 @@ export class BoxConfigService implements OnModuleInit {
         throw new BadRequestException('Box not found!');
       }
       //TODO:comment in
-      // const isVerified = checkIfMessageIsSigned(
-      //   signedMessage,
-      //   'Update Primes Mint',
-      //   authority,
-      // );
-      // if(!isVerified) throw new UnauthorizedException()
+      const isVerified = checkIfMessageIsSigned(
+        signedMessage,
+        'Update Primes Mint',
+        authority,
+      );
+      if (!isVerified) throw new UnauthorizedException();
       this.workers[boxIndex].box.boxState = BoxState.Removed;
       this.workers.splice(boxIndex, 1);
       const existingBox = await this.boxConfigRepo.findOne({
@@ -176,11 +177,10 @@ export class BoxConfigService implements OnModuleInit {
     const box = this.workers?.find((b) => b.box.boxId.toString() === boxId);
 
     if (!box) throw new NotFoundException('Given box not found!');
-    if (box.activeNft.nftId !== nftId)
-      throw new Error('Invalid NFT.Please try again!');
+    if (box.activeNft?.nftId !== nftId)
+      throw new Error('NFT expired, please try on next shuffle!');
     return box.placeBid(serializedTx);
   }
-
   checkBoxExistance(boxId: number) {
     return this.workers.find((w) => w.box.boxId === boxId);
   }
@@ -190,12 +190,12 @@ export class BoxConfigService implements OnModuleInit {
   }
   async deleteAllBoxes(signedMessage: string, authority: string) {
     //TODO:comment in
-    // const isVerified = checkIfMessageIsSigned(
-    //   signedMessage,
-    //   'Update Primes Mint',
-    //   authority,
-    // );
-    // if(!isVerified) throw new UnauthorizedException()
+    const isVerified = checkIfMessageIsSigned(
+      signedMessage,
+      'Update Primes Mint',
+      authority,
+    );
+    if (!isVerified) throw new UnauthorizedException();
     await this.boxConfigRepo.delete({});
 
     this.workers = [];
